VS 2017 (15.7) /analyze fixes

+++ DirectXTex/BC.h
@@ -170,7 +170,7 @@ struct D3DX_BC3
 //-------------------------------------------------------------------------------------
 #pragma warning(push)
 #pragma warning(disable : 4127)
-template <bool bRange> void OptimizeAlpha(float *pX, float *pY, const float *pPoints, size_t cSteps)
+template <bool bRange> void OptimizeAlpha(float *pX, float *pY, const float *pPoints, uint32_t cSteps)
 {
     static const float pC6[] = { 5.0f / 5.0f, 4.0f / 5.0f, 3.0f / 5.0f, 2.0f / 5.0f, 1.0f / 5.0f, 0.0f / 5.0f };
     static const float pD6[] = { 0.0f / 5.0f, 1.0f / 5.0f, 2.0f / 5.0f, 3.0f / 5.0f, 4.0f / 5.0f, 5.0f / 5.0f };
@@ -247,14 +247,13 @@ template <bool bRange> void OptimizeAlpha(float *pX, float *pY, const float *pPo
         {
             float fDot = (pPoints[iPoint] - fX) * fScale;
 
-            size_t iStep;
-
+            uint32_t iStep;
             if (fDot <= 0.0f)
                 iStep = ((6 == cSteps) && (pPoints[iPoint] <= fX * 0.5f)) ? 6 : 0;
             else if (fDot >= fSteps)
                 iStep = ((6 == cSteps) && (pPoints[iPoint] >= (fY + 1.0f) * 0.5f)) ? 7 : (cSteps - 1);
             else
-                iStep = static_cast<int32_t>(fDot + 0.5f);
+                iStep = uint32_t(fDot + 0.5f);
 
 
             if (iStep < cSteps)
